Transpiles primitive types
Also started writing the namespace, if it's present
