+++ src/main/scala/AvroOrder.scala
@@ -1,0 +1,6 @@
+package io.carrera.jsontoavroschema
+
+object AvroOrder extends Enumeration {
+  type AvroOrder = Value
+  val Ascending, Descending, Ignore = Value
+}
