+++ src/main/scala/AvroWriter.scala
@@ -3,13 +3,16 @@ package io.carrera.jsontoavroschema
 object AvroWriter {
 
   def toJson(record: AvroRecord): Either[TranspileError, ujson.Obj] = {
-    val result =
+    val result = {
       ujson.Obj(
         "type" -> "record",
         "name" -> record.name,
         "fields" -> record.fields.map(toJson)
       )
-    record.doc.map(d => result("doc") = d)
+    }
+
+    record.namespace.foreach(ns => result("namespace") = ns)
+    record.doc.foreach(d => result("doc") = d)
 
     Right(result)
   }
@@ -18,16 +21,11 @@ object AvroWriter {
     val result =
       ujson.Obj(
         "name" -> field.name,
-        "type" -> serializeType(field.`type`)
+        "type" -> AvroType.serialize(field.`type`)
       )
-    field.doc.map(d => result("doc") = d)
+    field.doc.foreach(d => result("doc") = d)
 
     result
   }
-
-  private def serializeType(avroType: AvroType) =
-    avroType match {
-      case _: AvroString => "string"
-    }
 
 }
